+++ lib/spid/configuration.rb
@@ -15,6 +15,8 @@ module Spid
     attr_accessor :certificate
     attr_accessor :attribute_service_name
     attr_accessor :default_relay_state_path
+    attr_accessor :acs_binding
+    attr_accessor :slo_binding
 
     # rubocop:disable Metrics/MethodLength
     def initialize
@@ -26,11 +28,13 @@ module Spid
       @slo_path                 = "/spid/slo"
       @digest_method            = Spid::SHA256
       @signature_method         = Spid::RSA_SHA256
-      @attribute_service_name   = attribute_service_name
+      @attribute_service_name   = nil
       @hostname                 = nil
       @private_key              = nil
       @certificate              = nil
       @default_relay_state_path = "/"
+      @acs_binding              = Spid::BINDINGS_HTTP_POST
+      @slo_binding              = Spid::BINDINGS_HTTP_REDIRECT
     end
     # rubocop:enable Metrics/MethodLength